+++ core/components/styles/nav.module.css
@@ -35,13 +35,6 @@
   font-family: "Montserrat", sans-serif;
 }
 
-.navbarRight {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  padding-right: 0.25rem;
-}
-
 .navbarLeft {
   display: flex;
   align-items: center;
@@ -99,11 +92,8 @@
     align-items: center;
   }
 
+  .navRightButton,
   .navbarLeft {
-    display: none;
-  }
-  
-  .navbarRight {
     display: none;
   }
