+++ core/components/styles/sidebar.module.css
@@ -46,7 +46,39 @@
 }
 
 .bottomButtons {
+  display: block;
+  margin-top: auto;
+}
+
+.logoutButton,
+.settingsButton {
+  align-items: center;
+  gap: 10px;
+  width: 100%;
+  padding: 10px;
+  border: none;
+  cursor: pointer;
+  color: white;
+  background: none;
+  text-align: left;
+
+  font-family: "Montserrat", sans-serif;
+}
+
+.settingsButton {
+  display: flex;
+  background-color: #59DBBC;
+  margin-top: 10px;
+}
+
+.logoutButton {
   display: none;
+  padding-left: 2rem;
+}
+
+.logoutButton:hover,
+.settingsButton:hover {
+  opacity: 0.8;
 }
 
 @media (max-width: 768px) {
@@ -75,38 +107,7 @@
     left: -250px;
   }
 
-  .bottomButtons {
-    display: block;
-    margin-top: auto;
-  }
-
-  .logoutButton,
-  .settingsButton {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    width: 100%;
-    padding: 10px;
-    border: none;
-    cursor: pointer;
-    color: white;
-    background: none;
-    text-align: left;
-
-    font-family: "Montserrat", sans-serif;
-  }
-
-  .settingsButton {
-    background-color: #59DBBC;
-    margin-top: 10px;
-  }
-
   .logoutButton {
-    padding-left: 2rem;
-  }
-
-  .logoutButton:hover,
-  .settingsButton:hover {
-    opacity: 0.8;
+    display: flex;
   }
 }
