+++ modules/lang/en/lang.json
@@ -36,7 +36,7 @@
   "Click To Stop Recording": "Click To Stop Recording",
   "Drag & Drop Or": "Drag & Drop Or",
   "Browse": "Browse",
-  "Supported Formats: MP3": "Supported Formats: MP3",
+  "Supported Formats: MP3. Max File Size: 50MB": "Supported Formats: MP3. Max File Size: 50MB-",
   "Upload Files": "Upload Files",
   "Upload": "Upload",
   "Success": "Success",
